+++ src/components/TreeInfo/AddTree.jsx
@@ -1,91 +1,255 @@
-import { useState, useRef } from "react";
+import { Button } from "@/components/ui/button"
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form"
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod"
+import {
+  Select,
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectLabel,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select"
+import { Input } from "@/components/ui/input"
+import {
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+} from "@/components/ui/accordion"
+import { Textarea } from "@/components/ui/textarea"
+import { toast } from "@/components/ui/use-toast"
+
+import TreeDrawer from "./TreeDrawer";
+import { NewTreeSchema } from "@/schema";
+
+import { useState, useEffect, useRef } from "react";
 import { addDoc, collection, serverTimestamp } from 'firebase/firestore';
 import { auth, db } from "../../utils/firebase";
 import * as displayText from "../../utils/displayText";
 import './TreeInfo.css'
 
-export default function AddTree({ mapCenter }) {
-    console.log(mapCenter)
-    const treeTypes = displayText.treeTypes
-    const accessMap = displayText.accessMap
-    
-    const [latitude, setLatitude] = useState(mapCenter.latitude);
-    const [longitude, setLongitude] = useState(mapCenter.longitude);
-    const [treeType, setTreeType] = useState('apple');
-    const [numTrees, setNumTrees] = useState(1);
-    const [access, setAccess] = useState('unknown');
-    const [notes, setNotes] = useState('');
-    
-    const treesCollectionRef = collection(db, "trees");
+export default function AddTreeNew({ isAddTreeVisible, setIsAddTreeVisible, draggablePosition }) {
+  const [key, setKey] = useState(+new Date())
 
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-        if (latitude === "" || longitude === "") return;
+  const form = useForm({
+    resolver: zodResolver(NewTreeSchema),
+    defaultValues: {
+      latitude: draggablePosition.lat,
+      longitude: draggablePosition.lng,
+      type: "apple",
+      treeCount: 1,
+      access: "unknown",
+      notes: ""
+    }
+  })
 
-        await addDoc(treesCollectionRef, {
-            description: newTree,
-            location: {
-                latitude: parseFloat(latitude),
-                longitude: parseFloat(longitude),
-            },
-            createdAt: serverTimestamp(),
-            userDisplayName: auth.currentUser.displayName,
-            userEmail: auth.currentUser.email,
-        });
+  const {
+    register,
+    handleSubmit,
+    setValue,
+    reset,
+    formState,
+    formState: { isSubmitting, isSubmitSuccessful }
+  } = form;
 
-        setNewTree("");
-        setLatitude("");
-        setLongitude("");
-    };
+  console.log("issubmit", isSubmitting)
+  console.log("issubmit successfull", isSubmitSuccessful)
 
-    return (
-        <div className="add-tree-menu">
-            {/* <Map 
-                mapSize='mini'
-                zoomSetting={18} 
-                mapCenter={[activeTree.geometry.coordinates[1],activeTree.geometry.coordinates[0]]}
-            /> */}
+  const treesCollectionRef = collection(db, "tree-features");
 
-            <p><b>New tree</b></p>
-            <form onSubmit={handleSubmit} className="new-tree-form">
-                <div>
-                    <label>Latitude: </label>
-                    <input type="number" value={latitude} onChange={(e) => setLatitude(e.target.value)} />
-                </div>
-                <div>
-                    <label>Longitude: </label>
-                    <input type="number" value={longitude} onChange={(e) => setLongitude(e.target.value)} />
-                </div>
-                <div>
-                    <label>Type: </label>
-                    <select value={treeType} onChange={(e) => setTreeType(e.target.value)}>
-                    {Object.keys(treeTypes).map((type) => (
-                        <option key={type} value={type}>
-                        {type}
-                        </option>
-                    ))}
-                    </select>
-                </div>
-                <div>
-                    <label>Number of trees: </label>
-                    <input type="number" value={numTrees} onChange={(e) => setNumTrees(e.target.value)} />
-                </div>
-                <div>
-                    <label>Access: </label>
-                    <select value={access} onChange={(e) => setAccess(e.target.value)}>
-                    {Object.keys(accessMap).map((key) => (
-                        <option key={key} value={key}>
-                        {accessMap[key].text}
-                        </option>
-                    ))}
-                    </select>
-                </div>
-                <div>
-                    <label>Notes: </label>
-                    <textarea value={notes} onChange={(e) => setNotes(e.target.value)} placeholder="Location details, tasting notes, access concerns" />
-                </div>
-                <button type="submit">Submit</button>
-            </form>
+  const onSubmit = async (data) => {
+    console.log('submitted!')
+
+    await addDoc(treesCollectionRef, {
+      location: {
+        latitude: data.latitude,
+        longitude: data.longitude,
+      },
+      type: data.type,
+      treeCount: data.treeCount,
+      access: data.access,
+      notes: data.notes,
+      createdAt: serverTimestamp(),
+      userDisplayName: auth.currentUser ? auth.currentUser.displayName : "no username",
+      userEmail: auth.currentUser ? auth.currentUser.email : null,
+      new: true
+  });
+
+    await new Promise((resolve) => setTimeout(resolve, 1000))
+
+    // TODO: if isSubmitSuccessful is true:
+    // hide drawer, display toast, remove active tree 
+
+    // toast({
+    //   title: "You submitted the following values:",
+    //   description: (
+    //     <pre className="mt-2 w-[340px] rounded-md bg-slate-950 p-4">
+    //       <code className="text-white">{JSON.stringify(data, null, 2)}</code>
+    //     </pre>
+    //   ),
+    // })
+  }
+
+  useEffect(() => {
+    // Watch for changes in draggablePosition and update form values accordingly
+    setValue('latitude', draggablePosition.lat);
+    setValue('longitude', draggablePosition.lng);
+  }, [draggablePosition, setValue]);
+
+  // reset form after submission
+  useEffect(() => {
+    if (formState.isSubmitSuccessful) {
+      setKey(+new Date())
+      reset();
+    }
+  }, [formState, reset]);
+
+  return (
+    <TreeDrawer
+    title="Add a new tree"
+    label="Drag the marker to adjust tree location"
+    open={isAddTreeVisible}
+    setOpen={setIsAddTreeVisible}
+    >
+    <Form {...form}>
+      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4 p-4 pb-0 pt-0">
+        <div className="space-y-2">
+        <FormField
+            control={form.control}
+            name="latitude"
+            render={({ field }) => (
+              <FormItem className="hidden">
+                <FormLabel>Latitude</FormLabel>
+                <FormControl>
+                  <Input {...field} type="number" placeholder="" />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        <FormField
+            control={form.control}
+            name="longitude"
+            render={({ field }) => (
+              <FormItem className="hidden">
+                <FormLabel>Longitude</FormLabel>
+                <FormControl>
+                  <Input {...field} type="number" placeholder="" />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="type"
+            render={({ field }) => (
+              <FormItem className="space-y-1">
+                <FormLabel>Type</FormLabel>
+                <Select onValueChange={field.onChange} defaultValue={field.value} key={key}>
+                <FormControl>
+                  <SelectTrigger>
+                    <SelectValue placeholder="Select a type of tree" />
+                  </SelectTrigger>
+                </FormControl>
+                <SelectContent>
+                  <SelectItem value="apple">Apple</SelectItem>
+                  <SelectItem value="fruit">Mixed</SelectItem>
+                  <SelectItem value="plum">Plum</SelectItem>
+                  <SelectItem value="cherry">Cherry</SelectItem>
+                  <SelectItem value="pear">Pear</SelectItem>
+                  <SelectItem value="crabapple">Crabapple</SelectItem>
+                  <SelectItem value="walnut">Walnut</SelectItem>
+                  <SelectItem value="other">Other</SelectItem>
+                </SelectContent>
+              </Select>
+              <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="treeCount"
+            render={({ field }) => (
+              <FormItem className="space-y-1">
+                <FormLabel>Number of trees</FormLabel>
+                <FormControl>
+                  <Input {...field} type="number" placeholder="" />
+                </FormControl>
+                <FormMessage />
+              <FormDescription>
+              </FormDescription>
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="access"
+            render={({ field }) => (
+              <FormItem className="space-y-1">
+                <FormLabel>Access</FormLabel>
+                <Select onValueChange={field.onChange} defaultValue={field.value} key={key}>
+                <FormControl>
+                  <SelectTrigger>
+                    <SelectValue placeholder="Select ownership and access status" />
+                  </SelectTrigger>
+                </FormControl>
+                <SelectContent>
+                  <SelectItem value="unknown">Unknown</SelectItem>
+                  <SelectItem value="public">Public</SelectItem>
+                  <SelectItem value="private-shared">Private - Shared</SelectItem>
+                  <SelectItem value="private">Private - Not Shared</SelectItem>
+                </SelectContent>
+              </Select>
+              <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="notes"
+            render={({ field }) => (
+              <FormItem className="space-y-1">
+                <Accordion type="single" collapsible>
+                  <AccordionItem value="item-1">
+                  <FormLabel><AccordionTrigger>Notes</AccordionTrigger></FormLabel>
+                    <AccordionContent>
+                      <FormControl>
+                        <Textarea {...field} className="resize-none" placeholder="Location details, tasting notes, access concerns..." />
+                      </FormControl>
+                    </AccordionContent>
+                  </AccordionItem>
+                </Accordion>
+              </FormItem>
+            )}
+          />
         </div>
-    );
+        <div className="button-row flex space-x-4">
+          <Button variant="outline" className="w-full"
+          onClick={(e) => {
+            e.preventDefault();
+            setKey(+new Date())
+            reset(undefined)
+          }}
+            >
+              Reset</Button>
+          <Button type="submit" className="w-full" disabled={isSubmitting}>
+            {isSubmitting ? "Submitting..." : "Submit"}
+          </Button>
+        </div>
+      </form>
+    </Form>
+    </TreeDrawer>
+    
+  )
 }
